# remove explicit return block that terminates rspec run