--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,7 +1,12 @@
-# 2.1.2
+## 2.1.3
+  - Fix spec early termination by removing an explicit return from a block
+
+## 2.1.2
   - Depend on logstash-core-plugin-api instead of logstash-core, removing the need to mass update plugins on major releases of logstash
-# 2.1.1
+
+## 2.1.1
   - New dependency requirements for logstash-core for the 5.0 release
+
 ## 2.1.0
  - Add caches for failed and successful lookups
  - Lower default timeout value