because of an explicit return from a block the spec suite was terminating earlier and not running all the tests

before this PR (10 tests are executed):

```
% bundle exec rspec -s 59466
Run options: exclude {:redis=>true, :socket=>true, :performance=>true, :couchdb=>true, :elasticsearch=>true, :elasticsearch_secure=>true, :export_cypher=>true, :integration=>true, :windows=>true}
.........

Finished in 0.899 seconds (files took 4.16 seconds to load)
10 examples, 0 failures

Randomized with seed 59466
```

after this pr (21 tests are executed):

```
% bundle exec rspec -s 59466
Run options: exclude {:redis=>true, :socket=>true, :performance=>true, :couchdb=>true, :elasticsearch=>true, :elasticsearch_secure=>true, :export_cypher=>true, :integration=>true, :windows=>true}
.....................

Finished in 1.58 seconds (files took 4.03 seconds to load)
21 examples, 0 failures

Randomized with seed 59466
```